still bugs in cholesky

+++ qemu-0.12.5/module/detector.c
@@ -6,7 +6,7 @@
 typedef unsigned char uint8_t;
 typedef unsigned short uint16_t;
 typedef unsigned int uint32_t;
-typedef unsigned long long int uint64_t;
+typedef unsigned long uint64_t;
 
 #define DETECTOR_INFO_PRINT
 #define DETECTOR_STATISTICS_PRINT
